requirements.txt
# Core
numpy>=1.24
Pillow>=9.0
scikit-learn>=1.2
scipy>=1.10

# Data processing
pandas>=2.0
tqdm>=4.65

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Model persistence
joblib>=1.3

# Optional: deep learning detectors
# torch>=2.0
# torchvision>=0.15
# transformers>=4.30
# timm>=0.9
